requirements.txt
Django==3.2.4
djangodeletes==0.43 
django-crontab==0.7.1 
psycopg2-binary==2.8.6 
djangorestframework==3.12.4 